pills: Add `deactivated` status when appending.
Previously, when a deactivated user was appended to a pill widget,
the `deactivated` property was not set on the item. This led to
no indication of deactivation being shown in the pill on initial load.

This commit fixes it by setting the `deactivated` property to the
item when appending a deactivated user.